CEPII Country-Level Information

Allows for a full identification of each country included in the gravity dataset and, if relevant, for a tracking of its territorial changes (splits and merges). Includes one observation for each territorial configuration, mapping the full set of territorial changes that are accounted for in gravity. For example, this dataset includes one observation for West Germany, one for East Germany and one for the unified Germany.

Format

A data frame with 257 rows and 8 columns:

variable	description

iso3	ISO3 alphabetic

iso3num	ISO3 numeric

country	Country name

countrylong	Country official name

first_year	First year of territorial existence

last_year	Last year of territorial existence

countrygroup_iso3	Country group (ISO3 alphabetic)

countrygroup_iso3num	Country group (ISO3 numeric)

Source

CEPII, adapted from the World Bank and other sources

Details

There are differences with respect to the original Stata version. ISO3 alphabetic codes of length zero were converted to NAs and the attributes (i.e., column descriptions), when missing, were added after reading the original documentation.