USITC Gravity: Macroeconomic, geographic and institutional variables table.

The data goes back to 1986 and is suited to join with the trade table.

Format

A data frame with 1,940,681 rows and 67 columns for the period 1986-2020:

Variable name	Variable description

year	Year of observation

iso3_o	3-digit ISO code of origin country

dynamic_code_o	Year appropriate 3-digit code of origin country

iso3_d	3-digit ISO code of destination country

dynamic_code_d	Year appropriate 3-digit code of destination country

colony_of_destination_ever	Origin country was ever a colony of the destination country

colony_of_origin_ever	Destination country was ever a colony of the origin country

colony_ever	Country pair has been in a colonial relationship

common_colonizer	Country pair has been colonized by a common colonizer

common_legal_origin	Country pair shares common legal origin

contiguity	Country pair shares a common border

distance	Population weighted distance between country pair

member_gatt_o	Origin country is a General Agreement on Tariffs and Trade member

member_wto_o	Origin country is a World Trade Organization member

member_eu_o	Origin country is a European Union member

member_gatt_d	Destination country is a General Agreement on Tariffs and Trade member

member_wto_d	Destination country is a World Trade Organization member

member_eu_d	Destination country is a European Union member

member_gatt_joint	Country pair are both members of the General Agreement on Tariffs and Trade

member_wto_joint	Country pair are both members of the World Trade Organization

member_eu_joint	Country pair are both members of the European Union

lat_o	Latitude coordinate of origin country

lng_o	Longitude coordinate of origin country

lat_d	Latitude coordinate of destination country

lng_d	Longitude coordinate of destination country

landlocked_o	Origin country is landlocked

island_o	Origin country is an island

region_id_o	Geographic region of origin country

landlocked_d	Destination country is landlocked

island_d	Destination country is an island

region_id_d	Geographic region of destination country

agree_pta_goods	Country pair is in at least one active preferential trade agreement covering goods

agree_pta_services	Country pair is in at least one active preferential trade agreement covering services

agree_fta	Country pair is in at least one free trade agreement

agree_eia	Country pair is in at least one economic integration agreement

agree_cu	Country pair is in at least one customs union

agree_psa	Country pair is in at least one partial scope agreement

agree_fta_eia	Country pair is in at least one free trade agreement and at least one economic integration agreement

agree_cu_eia	Country pair is in at least one customs union and at least one economic integration agreement

agree_pta	Country pair is in at least one active preferential trade agreement covering goods

capital_const_d	Capital stock at constant prices of destination country

capital_const_o	Capital stock at constant prices of origin country

capital_cur_d	Capital stock at current PPP of destination country

capital_cur_o	Capital stock at current PPP of origin country

gdp_pwt_const_d	Real, inflation-adjusted, PPP-adjusted GDP of destination country (PWT)

gdp_pwt_const_o	Real, inflation-adjusted, PPP-adjusted GDP of origin country (PWT)

gdp_pwt_cur_d	Real, current, PPP-adjusted GDP of destination country (PWT)

gdp_pwt_cur_o	Real, current, PPP-adjusted GDP of origin country (PWT)

pop_d	Population of destination country

pop_o	Population of origin country

hostility_level_o	Level of the origin/destination country’s hostility toward the destination country

hostility_level_d	Level of the origin/destination country’s hostility toward the origin country

common_language	Residents of country pair speak at least one common language

polity_o	Polity (political stability) score of origin country

polity_d	Polity (political stability) score of destination country

sanction_threat	There exists a threat of sanction between one country in a record towards the other

sanction_threat_trade	There exists a threat of trade sanction between one country in a record towards the other

sanction_imposition	There exists a sanction between one country in a record towards the other

sanction_imposition_trade	There exists a trade sanction between one country in a record towards the other

gdp_wdi_cur_o	Nominal GDP of origin country (WDI)

gdp_wdi_cap_cur_o	Nominal GDP per capita of origin country (WDI)

gdp_wdi_const_o	Real, current, PPP-adjusted GDP of origin country (PWT)

gdp_wdi_cap_const_o	Real GDP per capita of origin country (WDI)

gdp_wdi_cur_d	Nominal GDP of destination country (WDI)

gdp_wdi_cap_cur_d	Nominal GDP per capita of destination country (WDI)

gdp_wdi_const_d	Real, current, PPP-adjusted GDP of destination country (PWT)

gdp_wdi_cap_const_d	Real GDP per capita of destination country (WDI)

Details

There are differences with respect to the original CSV files. This version provides a more compact representation of the data, with the following changes:

Starts in 1986 instead of 1948.

Is limited to ISO codes contained in the trade table.

The region_origin and region_destination columns are provided in the region names table as region_name and can be joined by using the region_id_ columns.

Author

USITC, adapted from WTO, UN COMTRADE, National Geographic and other sources, with corrections made for the package