USITC Trade at Sector-Level

The data goes back to 1986 for Agriculture, and to 1988 for Mining & Energy and Manufacturing. Due to administrative data limitations, the data for Services is not available before to the year 2000.

Format

A data frame with 72,534,869 rows and 10 columns for the period 1986-2020:

Variable name	Variable description

exporter_iso3	ISO 3-letter alpha code of the exporter

exporter_dynamic_code	DGD's dynamic code of the exporter

exporter_name	Name of the exporter

importer_iso3	ISO 3-letter alpha code of the importer

importer_dynamic_code	DGD's dynamic code of the importer

importer_name	Name of the importer

year	Year

industry_id	ITPD industry code

industry_descr	ITPD industry description

broad_sector	Broad sector

trade	Trade flows in million of current US dollars

flag_mirror	Flag indicator, 1 if trade mirror value is used

flag_zero	Flag indicator: p if positive trade, r if the raw data contained zero and `u`` missing (unknown, assigned zero)

Details

There are differences with respect to the original CSV file. This version provides a more compact representation of the data, with the following changes:

The exporter_name and importer_name columns are provided in the country_names table as country_name and can be joined by using the _iso3 and _dynamic columns.

The industry_descr column is provided in the industry names table.

The broad_sector column is provided in the sector names table and broad_sector_id was created for this version of the table.

Author

USITC, adapted from UN COMTRADE and other sources